# frontend/apps/web/public/specs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "TraceRTM Backend API - Requirements Traceability Management System with AI-powered agents\n\nTraceRTM is a comprehensive requirements traceability management system that helps teams\ntrack relationships between requirements, features, and tasks throughout the development lifecycle.\n\n## Features\n- Project and Item Management\n- Link Management for traceability\n- AI Agent Coordination\n- Graph-based relationship analysis\n- Full-text search with pgvector\n- Real-time updates via WebSocket and NATS\n\n## Authentication\nCurrently supports WorkOS AuthKit. Set Authorization header with Bearer token.",
        "title": "TraceRTM API",
        "contact": {
            "name": "TraceRTM Support",
            "url": "https://github.com/kooshapari/tracertm-backend",
            "email": "support@tracertm.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/auth/logout": {
            "post": {
                "description": "Logout and revoke session",
                "tags": [
                    "auth"
                ],
                "summary": "User logout",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "description": "Get the current authenticated user information",
                "tags": [
                    "auth"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "Current user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "Refresh expired or expiring authentication token",
                "tags": [
                    "auth"
                ],
                "summary": "Refresh authentication token",
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns information about the currently authenticated user",
                "tags": [
                    "auth"
                ],
                "summary": "Get current user information",
                "responses": {
                    "200": {
                        "description": "Current user information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth.User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handlers.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "description": "Enter your bearer token in the format: Bearer {token}",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
            "auth.User": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "metadata": {
                        "type": "object",
                        "additionalProperties": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "projectID": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    }
                }
            },
...
            "handlers.AuthResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/auth.User"
                    }
                }
            },
...
            "handlers.ErrorResponse": {
                "description": "Error response with error message",
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}